remove internal package. These codes are moved into cmd/rrh/commands by using cobra
